requirements.txt
aiofiles==24.1.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
boto3==1.40.59
botocore==1.40.59
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cryptography==46.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
mailbridge==2.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
asyncpg>=0.30.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.5
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.5.1
rsa==4.9.1
s3transfer==0.14.0
sentry-sdk==2.32.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
structlog==24.1.0
text-unidecode==1.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3